+++ jdkringbuffer-core/src/main/java/org/jdkstack/jdkringbuffer/core/spsc/version2/AbstractSpscBlockingQueueV2.java
@@ -1,10 +1,6 @@
 package org.jdkstack.jdkringbuffer.core.spsc.version2;
 
-import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.TimeUnit;
-import org.jdkstack.jdkringbuffer.api.RingBufferBlockingQueue;
-import org.jdkstack.jdkringbuffer.core.AbstractBlockingQueue;
-import org.jdkstack.jdkringbuffer.core.Constants;
+import org.jdkstack.jdkringbuffer.core.AbstractLockBlockingQueueV2;
 import org.jdkstack.jdkringbuffer.core.Entry;
 
 /**
@@ -15,21 +11,7 @@ import org.jdkstack.jdkringbuffer.core.Entry;
  * @author admin
  * @param <E> e .
  */
-public abstract class AbstractSpscBlockingQueueV2<E> extends AbstractBlockingQueue<E>
-    implements BlockingQueue<E>, RingBufferBlockingQueue {
-  /** 环形数组. */
-  private final Entry<E>[] buffer;
-
-  /**
-   * This is a method description.
-   *
-   * <p>Another description after blank line.
-   *
-   * @author admin
-   */
-  protected AbstractSpscBlockingQueueV2() {
-    this(Constants.CAPACITY);
-  }
+public abstract class AbstractSpscBlockingQueueV2<E> extends AbstractLockBlockingQueueV2<E> {
 
   /**
    * This is a method description.
@@ -39,13 +21,8 @@ public abstract class AbstractSpscBlockingQueueV2<E> extends AbstractBlockingQue
    * @author admin
    * @param capacity e.
    */
-  @SuppressWarnings("unchecked")
   protected AbstractSpscBlockingQueueV2(final int capacity) {
-    super(capacity, capacity - 1);
-    buffer = new Entry[capacity];
-    for (int i = 0; i < capacity; i++) {
-      buffer[i] = new Entry<>(i);
-    }
+    super(capacity);
   }
 
   /**
@@ -95,33 +72,5 @@ public abstract class AbstractSpscBlockingQueueV2<E> extends AbstractBlockingQue
       cell.setSeq(headSeq + index + 1);
     }
     return e;
-  }
-
-  /**
-   * This is a method description.
-   *
-   * <p>Another description after blank line.
-   *
-   * @author admin
-   * @return E e.
-   */
-  @Override
-  public E poll(final long timeout, final TimeUnit unit) {
-    throw new UnsupportedOperationException("未实现.");
-  }
-
-  @Override
-  public final E peek() {
-    return buffer[head.get() & index].getEntry();
-  }
-
-  @Override
-  public final int size() {
-    return Math.max(tail.get() - head.get(), 0);
-  }
-
-  @Override
-  public final boolean isEmpty() {
-    return head.get() == tail.get();
   }
 }
